# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Service culturel - Plateforme web API",
    "description": "API de gestion des inscriptions scolaires aux événements de l'Opéra Orchestre National Montpellier Occitanie.",
    "version": "1.6.7",
    "contact": {
      "name": "Service culturel - Plateforme web",
      "url": "https://github.com/vincbct34/service-culturel-plateforme-web"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Serveur courant"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token JWT (header Authorization: Bearer <token>)"
      },
      "CronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET token pour les tâches planifiées"
      },
      "CSRFToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Jeton CSRF requis pour POST/PUT/PATCH/DELETE authentifiés"
      }
    },
    "schemas": {
...
      "PublicCategory": {
        "type": "string",
        "enum": [
          "CRECHE",
          "MATERNELLE",
          "ELEMENTAIRE",
          "COLLEGE",
          "LYCEE",
          "SUPERIEUR",
          "CONSERVATOIRE",
          "ASSOCIATION",
          "PERISCOLAIRE",
          "PUBLICS_EMPECHES",
          "AUTRE"
        ]
      },
      "SchoolGrade": {
        "type": "string",
        "enum": [
          "PS",
          "MS",
          "GS",
          "CP",
          "CE1",
          "CE2",
          "CM1",
          "CM2",
          "SIXIEME",
          "CINQUIEME",
          "QUATRIEME",
          "TROISIEME",
          "SECONDE",
          "PREMIERE",
          "TERMINALE"
        ]
      },
      "AgeRange": {
        "type": "string",
        "enum": ["AGE_0_3", "AGE_3_6", "AGE_6_11", "AGE_11_15", "AGE_15_18", "AGE_18_PLUS"]
      },
...
      "Accessibility": {
        "type": "string",
        "enum": ["VISUAL", "AUDITORY", "MOTOR", "PSYCHIC", "NEUROATYPICAL", "OTHER"]
      },
      "RegistrationStatus": {
        "type": "string",
        "enum": ["PENDING", "CONFIRMED", "CANCELLED", "REJECTED", "ATTENDED", "NO_SHOW"]
      },
...
      "Disability": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Accessibility"
          },
          "count": {
            "type": "integer",
            "minimum": 1
          },
          "details": {
            "type": "string"
          }
        },
        "required": ["type", "count"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "Registration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "institution_id": {
            "type": "string"
          },
          "event_id": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "booked_seats": {
            "type": "integer"
          },
          "caretaker_count": {
            "type": "integer"
          },
          "aesh_count": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "manager_first_name": {
            "type": "string",
            "nullable": true
          },
          "manager_last_name": {
            "type": "string",
            "nullable": true
          },
          "manager_email": {
            "type": "string",
            "nullable": true
          },
          "manager_phone_number": {
            "type": "string",
            "nullable": true
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "was_present_comment": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCategory"
            }
          },
          "grades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchoolGrade"
            }
          },
          "age_ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgeRange"
            }
          },
          "want_formation": {
            "type": "boolean"
          },
          "want_preparation": {
            "type": "boolean"
          },
          "disabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Disability"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCategory"
            }
          },
          "grades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchoolGrade"
            }
          },
          "age_ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgeRange"
            }
          },
          "students_count": {
            "type": "integer"
          },
          "disabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Disability"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/institutions/{id}/registrations": {
      "get": {
        "tags": ["Institutions"],
        "summary": "Inscriptions d'une institution",
        "description": "Liste les inscriptions liées à une institution avec événement, utilisateur et handicaps.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "0 = pas de pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Inscriptions de l'institution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "registrations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registration"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/institutions/{id}/history": {
      "get": {
        "tags": ["Institutions"],
        "summary": "Historique d'une institution",
        "description": "Retourne l'historique analytics complet d'une institution (taux de présence, confirmations, no-shows, etc.). Admin uniquement. Résultat mis en cache.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detailed",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historique de l'institution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "institution": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "history": {
                      "type": "object",
                      "properties": {
                        "totalRegistrations": {
                          "type": "integer"
                        },
                        "confirmedCount": {
                          "type": "integer"
                        },
                        "attendedCount": {
                          "type": "integer"
                        },
                        "noShowCount": {
                          "type": "integer"
                        },
                        "cancelledCount": {
                          "type": "integer"
                        },
                        "attendanceRate": {
                          "type": "number"
                        },
                        "confirmationRate": {
                          "type": "number"
                        },
                        "summary": {
                          "type": "string"
                        },
                        "health": {
                          "type": "object"
                        },
                        "report": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{groupId}": {
      "patch": {
        "tags": ["Groups"],
        "summary": "Modifier un groupe",
        "description": "Met à jour un groupe. Propriétaire ou admin uniquement.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "nullable": true
                  },
                  "category": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PublicCategory"
                    }
                  },
                  "grades": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SchoolGrade"
                    }
                  },
                  "age_ranges": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AgeRange"
                    }
                  },
                  "students_count": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "disabilities": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Disability"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Groupe mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "group": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Groups"],
        "summary": "Supprimer un groupe",
        "description": "Supprime un groupe. Propriétaire ou admin.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groupe supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Groupe supprimé avec succès"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}